Add drupal entities:
Added the following entities:
DocFile
Doc
Link

Also set up creation of needed taxonomy vocab and terms on installation of module
(and delete on uninstallation)

Making a Doc entity sets the DocFile docProcessed field to processed.
Processing the number of links stored in the fromDoc entity will set the DocFile field of it to processed.